+++ lib/events/puppet_show.hpp
@@ -10,7 +10,7 @@ class PuppetShow: public InfantEvent {
 
     public:
         PuppetShow();
-        PuppetShow(int id, std::string category, std::string name, int n_responsible, std::vector<int> capacity, std::vector<int> prices,
+        PuppetShow(int id, std::string category, std::string name, int n_responsible, std::vector<int> capacity, std::vector<float> prices,
             std::vector<int> theaterSchedules);
 
         std::vector<int> get_schedules(){return theaterSchedules;}
