+++ src/events/consert.cpp
@@ -2,7 +2,7 @@
 
 Consert::Consert(){}
 
-Consert::Consert(int id, std::string category, std::string  name, int  responsible, std::vector<int>  capacity, std::vector<int>  prices,
+Consert::Consert(int id, std::string category, std::string  name, int  responsible, std::vector<int>  capacity, std::vector<float>  prices,
     int  elderAmount, int gateOpening, std::vector<std::string> artists){
 
     this->id =  id;
